requirements.txt
ansible-base==2.10.8
apache-libcloud==3.2.0
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.5.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==41.0.3
cupshelpers==1.0
daphne==4.0.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.0.3
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.6.1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==22.10.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
load-dotenv==0.1.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.5
mutagen==1.45.1
myenv==201902.7
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.0.0
platformdirs==2.5.1
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
pymongo==4.7.3
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2021.3
pywinrm==0.3.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.0
reportlab==3.6.8
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
selinux==3.3
service-identity==23.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
systemd-python==234
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2021.5
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
websockets==9.1
xdg==5
xkit==0.0.0
xmltodict==0.12.0
yt-dlp==2022.4.8
zipp==1.0.0
zope.interface==6.0
